requirements.txt
click==8.1.7
colorama==0.4.6
Flask==3.1.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
setuptools
gspread==6.1.0
google-auth==2.28.1
